acd-aidev/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
scipy>=1.9.0
joblib==1.4.2
matplotlib==3.9.2
seaborn==0.13.2
streamlit==1.39.0
paho-mqtt==2.1.0
requests==2.32.3
python-dotenv==1.0.1
